+++ src/index.ts
@@ -299,6 +299,32 @@ export {
   installmentsClientPresenceSchema,
   installmentsClientValidationSchema,
 } from "./installments/clients/validate-client/validate-client.js";
+export type { InstallmentsOrderPayment } from "./installments/orders/check-order-paid/check-order-paid.js";
+export { installmentsOrderPaymentSchema } from "./installments/orders/check-order-paid/check-order-paid.js";
+export type {
+  CreateInstallmentsOrderInput,
+  InstallmentsOrderInvoiceInput,
+  InstallmentsProductInput,
+  InstallmentsProgramInput,
+  NewInstallmentsOrder,
+} from "./installments/orders/create-order/create-order.js";
+export { newInstallmentsOrderSchema } from "./installments/orders/create-order/create-order.js";
+export type {
+  InstallmentsOrderData,
+  InstallmentsOrderReverse,
+} from "./installments/orders/models/installments-order-data.js";
+export {
+  installmentsOrderDataSchema,
+  installmentsOrderReverseSchema,
+} from "./installments/orders/models/installments-order-data.js";
+export type { InstallmentsOrderState } from "./installments/orders/models/installments-order-state.js";
+export { installmentsOrderStateSchema } from "./installments/orders/models/installments-order-state.js";
+export type {
+  InstallmentsOrderReturn,
+  ReturnInstallmentsOrderInput,
+} from "./installments/orders/return-order/return-order.js";
+export { installmentsOrderReturnSchema } from "./installments/orders/return-order/return-order.js";
+export type { InstallmentsOrderIdentifierInput } from "./installments/orders/shared/order-identifier.js";
 export { MonobankPersonalClient } from "./personal/client/monobank-personal-client.js";
 export type { MonobankPersonalClientOptions } from "./personal/client/monobank-personal-client-options.js";
 export type { ClientInfo } from "./personal/client-info/get-info/get-info.js";